fix: Input fields no longer lose values
Input fields lost their values if they were last in grades/add because the component was reloaded to add or remove the button to remove the specific grade input field.

The remove button was moved to the parent page out of the component and is now located next to the button to add new result input fields.